added the ability to exit at any time

+++ python_code.py
@@ -22,6 +22,11 @@ db = sqlite3.connect('animals')
 cursor = db.cursor()
 
 # functions
+def exit_check(check):
+    """checks if a input is exit"""
+    if check.lower() == "exit":
+        return True
+
 def ranking_help(value):
     get_stuff(f'SELECT ANIMALS.animal_id, ANIMALS.animal_name, ANIMALS.scientific_name, RANKING.could_i_take_it AS fight_rating, INFO."group" AS animal_group FROM ANIMALS JOIN RANKING ON ANIMALS.could_i_take_it_in_a_fight = RANKING.rank_id JOIN INFO ON ANIMALS.animal_info = INFO.info_id where could_i_take_it_in_a_fight = {value};')
 
@@ -36,32 +41,44 @@ def get_stuff(query):
 def ranking_nums():
     print(f'{y}from 81 being death, to 85 being easy as hell, {reset}')
     print(f'{y}81, 82, 83, 84, 85,{reset}')
+
 def group_names():
     '''prints group names'''
-    print(f'{y}86	Reptile,\n 87	Mammal,\n 88	Insect,\n 89	Arachnid,\n 90	Amphibian,\n 91	Bird,\n 92	Mollusk,\n 93	Cockroach,\n 94	Myriapod{reset}')
+    print(f'{y} 86	Reptile,\n 87	Mammal,\n 88	Insect,\n 89	Arachnid,\n 90	Amphibian,\n 91	Bird,\n 92	Mollusk,\n 93	Cockroach,\n 94	Myriapod{reset}')
 
 def add_animals():
     '''adds an animal'''
     animal_name = input(f'{bb}animal name: {reset}')
+    if exit_check(animal_name):
+        return
     scientific_name = input(f'{bb}scientific name: {reset}')
+    if exit_check(scientific_name):
+        return
     group_names()
-    group = int(input(f'{bb}group: {reset}'))
-    ranking = int(input(f'{bb}ranking, 81 to 85, 81 is death: {reset}'))
-    qurey = f"INSERT INTO ANIMALS (animal_name, scientific_name, animal_info, could_i_take_it_in_a_fight) VALUES('{animal_name.title()}', '{scientific_name.title()}', '{group}', '{ranking}');"
+    group = input(f'{bb}group: {reset}')
+    if exit_check(group):
+        return
+    ranking = input(f'{bb}ranking, 81 to 85, 81 is death: {reset}')
+    if exit_check(ranking):
+        return
+    qurey = f"INSERT INTO ANIMALS (animal_name, scientific_name, animal_info, could_i_take_it_in_a_fight) VALUES('{animal_name.title()}', '{scientific_name.title()}', '{int(group)}', '{int(ranking)}');"
     cursor.execute(qurey,)
 
 def select_by_ranking():
-    '''selects * by the ranking'''
-    search = int(input(f'{y}1. im COOKED \n2. 1 percent chance I dont lose \n3. ill probably lose \n4. I definitely have a chance \n5. Im whooping their ASS \n{bb}what ranking? \n{reset}'))    
-    if search == 1:
+    '''select * by the ranking'''
+    search = input(f'{y}1. im COOKED \n2. 1 percent chance I dont lose \n3. ill probably lose \n4. I definitely have a chance \n5. Im whooping their ASS \n{bb}what ranking? \n{reset}')    
+    if exit_check(search):
+        return
+    while run.isnumeric() or not run.isnumeric():
+    if int(search) == 1:
         ranking_help(81)
-    elif search == 2:
+    elif int(search) == 2:
         ranking_help(82)
-    elif search == 3:
+    elif int(search) == 3:
         ranking_help(83)
-    elif search == 4:
+    elif int(search) == 4:
         ranking_help(84)
-    elif search == 5:
+    elif int(search) == 5:
         ranking_help(85)
 
 def select_all():
@@ -71,28 +88,41 @@ def select_all():
 def search_animal():
     '''search for a specific animal'''
     animal = input(f'{bb}what animal would you like to search for? {reset}').title()
+    if exit_check(animal):
+        return
     get_stuff(f"select * from ANIMALS where animal_name = '{animal}';")
 
 def kill_an_animal():
     animal = input(f'{bb}what animal would you like to kill? {reset}').title()
+    if exit_check(animal):
+        return
     query = f'Delete from animals where animal_name = {animal.title()};'
     cursor.execute(query, )
 
 def edit_an_animal():
     print(f'{br}DO NOT CHANGE ANIMAL idS!!!!!!!!!!!!!!!{reset}')
     change = input(f'{bb}what animal do you want to change? {reset}')
+    if exit_check(change):
+        return
     print(f'{y}columms are; \nanimal_name\nscientific_name\ncould_i_take_it_in_a_fight\nanimal_info(group){reset}')
     columm = input(f'{bb}what columm do you want to edit? {reset}')
+    if exit_check(columm):
+        return
     if columm == 'could_i_take_it_in_a_fight':
         ranking_nums()
     if columm == 'group':
         group_names()
     changes = input(f"{bb}what do you want to change it to? {reset}")
+    if exit_check(changes):
+        return
     query = f"UPDATE animals SET {columm} = '{changes}' WHERE animal_name = '{change}';"
     cursor.execute(query,)
 
 def advanced_():
-    cursor.execute(input(f'{bb}Write your own query: {reset}'))
+    Qss = input(f'{bb}Write your own query: {reset}')
+    if exit_check(Qss):
+        return
+    cursor.execute(input(Qss))
     results = cursor.fetchall()
     for number in results:
         print(f"     {number}")
@@ -100,7 +130,7 @@ def advanced_():
 # main code
 
 print(options)
-run = input('what do you want to do? (enter a number) ')
+run = input(f'{bb}what do you want to do? (enter a number) {reset}')
 while run.isnumeric() or not run.isnumeric():
     
     if   run == '1':
@@ -123,4 +153,4 @@ while run.isnumeric() or not run.isnumeric():
         print(f'{r}no{reset}')
     
     print(options)
-    run = input('what do you want to do? ')
+    run = input(f'{bb}what do you want to do? {reset}')
